fix: Installer script for powershell
Changed to work with cross-platform PowerShell in addition to
Windows PowerShell.

Added a structure to facilitate platform branching in Installer.
This allows for finer specification of supported platforms.

Made changes to allow the use of GitHub access tokens,
since GitHub Rest API was being used without tokens inside scripts.
This avoids the impact of access restrictions based on IP address.